1.安裝npm  2.新增登入介面 新增admin, manager,與user 進行測試  3.更新兩個views： players.index 和teams.index  4.修改route經過 auth 的 middleware (即必須登入帳號才可以使用)  5.新增表單、儲存資料、刪除資料路由只能讓admin使用  6.編輯表單只允許 admin或manager使用

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PlayersController;
 use App\Http\Controllers\TeamsController;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,6 +15,8 @@ use App\Http\Controllers\TeamsController;
 | contains the "web" middleware group. Now create something great!
 |
 */
+Route::middleware(['auth'])->group(function () {
+
 Route::get('/',function(){
         return redirect('players');
 });
@@ -30,29 +33,29 @@ Route::get('players/{id}', [PlayersController::class, 'show'])->where('id', '[0-
 // 修改單一選手資料
 Route::get('players/{id}/edit', [PlayersController::class, 'edit'])->where('id', '[0-9]+')->name('players.edit');
 // 刪除單一選手資料
-Route::delete('players/delete/{id}', [PlayersController::class, 'destroy'])->where('id', '[0-9]+')->name('players.destroy');
+Route::delete('players/delete/{id}', [PlayersController::class, 'destroy'])->where('id', '[0-9]+')->name('players.destroy')->middleware('can:admin');
 // 新增選手表單
-Route::get('players/create', [PlayersController::class, 'create'])->name('players.create');
+Route::get('players/create', [PlayersController::class, 'create'])->name('players.create')->middleware('can:admin');
 // 修改選手表單
 Route::get('players/{id}/edit', [PlayersController::class, 'edit'])->where('id', '[0-9]+')->name('players.edit');
 // 修改選手資料
 Route::patch('players/update/{id}', [PlayersController::class, 'update'])->where('id', '[0-9]+')->name('players.update');
 // 儲存新選手資料
-Route::post('players/store', [PlayersController::class, 'store'])->where('id', '[0-9]+')->name('players.store');
+Route::post('players/store', [PlayersController::class, 'store'])->where('id', '[0-9]+')->name('players.store')->middleware('can:admin');
 
 Route::get('teams',[TeamsController::class, 'index'])->name('teams.index');
 Route::get('teams/{id}', [TeamsController::class, 'show'])->where('id', '[0-9]+')->name('teams.show');
 Route::get('teams/{id}/edit', [TeamsController::class, 'edit'])->where('id', '[0-9]+')->name('teams.edit');
 // 刪除單一戰隊及旗下選手資料
-Route::delete('teams/delete/{id}', [TeamsController::class, 'destroy'])->where('id', '[0-9]+')->name('teams.destroy');
+Route::delete('teams/delete/{id}', [TeamsController::class, 'destroy'])->where('id', '[0-9]+')->name('teams.destroy')->middleware('can:admin');
 // 新增戰隊表單
-Route::get('teams/create', [TeamsController::class, 'create'])->name('teams.create');
+Route::get('teams/create', [TeamsController::class, 'create'])->name('teams.create')->middleware('can:admin');
 // 修改戰隊表單
 Route::get('teams/{id}/edit', [TeamsController::class, 'edit'])->where('id', '[0-9]+')->name('teams.edit');
 // 修改戰隊資料
 Route::patch('teams/update/{id}', [TeamsController::class, 'update'])->where('id', '[0-9]+')->name('teams.update');
 // 儲存新戰隊資料
-Route::post('teams/store', [TeamsController::class, 'store'])->name('teams.store');
+Route::post('teams/store', [TeamsController::class, 'store'])->name('teams.store')->middleware('can:admin');
 
 // 顯示賽區第一
 Route::get('teams/rank1', [TeamsController::class, 'rank1'])->name('teams.rank1');
@@ -62,3 +65,8 @@ Route::get('teams/rank2', [TeamsController::class, 'rank2'])->name('teams.rank2'
 Route::get('teams/rank3', [TeamsController::class, 'rank3'])->name('teams.rank3');
 // 顯示賽區第四
 Route::get('teams/rank4', [TeamsController::class, 'rank4'])->name('teams.rank4');
+});
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
